修改接口

+++ me.cqp.luohuaming.UnraidMonitor.PublicInfos/Drawing/Items/DrawingItem_Text.cs
@@ -12,7 +12,7 @@ namespace me.cqp.luohuaming.UnraidMonitor.PublicInfos.Drawing.Items
 
         public bool IsBold { get; set; }
 
-        public override (SKPoint endPoint, double width, double height) Draw(Painting painting, SKPoint startPoint, double desireWidth, double desireHeight, DrawingStyle.Theme theme, DrawingStyle.Colors palette)
+        public override (SKPoint endPoint, float width, float height) Draw(Painting painting, SKPoint startPoint, double desireWidth, double desireHeight, DrawingStyle.Theme theme, DrawingStyle.Colors palette)
         {
             return base.Draw(painting, startPoint, desireWidth, desireHeight, theme, palette);
         }
